+++ cppss/vec.h
@@ -27,7 +27,7 @@ class vettore{
 		double modulo() const {return sqrt(m_x*m_x+m_y*m_y+m_z*m_z);}
 		friend std::ostream& operator<<(std::ostream &stream, const vettore c);
 		void sfToCar();
-		float angolo(const vettore &c);
+		double angolo(const vettore &c);
 };
 std::ostream& operator<<(std::ostream &stream, const vettore c);
 //float angolo(const vettore &b, vettore &c);
